requirements.txt
google-generativeai
python-dotenv
pymisp
selenium
dnstwist[full]
